clean build warning on windows

+++ lib/context.c
@@ -598,7 +598,8 @@ lws_create_context(struct lws_context_creation_info *info)
 
 	if (info->server_string) {
 		context->server_string = info->server_string;
-		context->server_string_len = strlen(context->server_string);
+		context->server_string_len = (short)
+				strlen(context->server_string);
 	} else {
 		context->server_string = "libwebsockets";
 		context->server_string_len = 13;
